requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
mccabe==0.7.0
packaging==23.2
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1